employees/models.py
from django.db import models
from mptt.models import MPTTModel, TreeForeignKey


class Subdivision(MPTTModel):
    name = models.CharField(max_length=255, unique=True)
    parent = TreeForeignKey('self', on_delete=models.CASCADE, null=True, blank=True, related_name='children')

    class MPTTMeta:
        order_insertion_by = ['name']

    def __str__(self):
        return self.name


class Employee(models.Model):
    full_name = models.CharField(max_length=255)
    position = models.CharField(max_length=100)
    hire_date = models.DateField()
    salary = models.DecimalField(max_digits=10, decimal_places=2)
    subdivision = models.ForeignKey(Subdivision, on_delete=models.CASCADE, related_name='employees')

    def __str__(self):
        return self.full_name
