requirements.txt
asgiref==3.8.1
Django==5.1.1
django-js-asset==2.2.0
django-mptt==0.16.0
mimesis==18.0.0
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.5.1
tzdata==2024.2
